save21

+++ player.c
@@ -21,12 +21,12 @@ int	move(t_frame *frame, int x, int y)
 	player->y += y;
 	if (player->x < 0)
 		player->x = 0;
-	else if (player->x >= frame->width)
-		player->x = frame->width - 1;
+	else if (player->x >= (int)(frame->width))
+		player->x = (int)(frame->width) - 1;
 	else if (player->y < 0)
 		player->y = 0;
-	else if (player->y >= frame->height)
-		player->y = frame->height - 1;
+	else if (player->y >= (int)(frame->height))
+		player->y = (int)(frame->height) - 1;
 	else if (frame->list[player->y][player->x] == WALL)
 	{
 		player->x -= x;
@@ -39,7 +39,6 @@ int	move(t_frame *frame, int x, int y)
 
 void	action(t_frame *frame)
 {
-
 	frame->player->move_count ++;
 	ft_putstr("Move Count : ", STDOUT);
 	ft_putstrln(ft_itoa(frame->player->move_count), STDOUT);
